requirements.txt
dnslib
pycrypto
colorama
